WebSockets: do not report connection errors as handshake failures
We are assuming that when not switching protocols after a request is
because the server didn't accept the handshake, but we can fail even
earlier if there's a connection error. In libsoup 2 we could check if
it was a connection error by checking the message status was
SOUP_STATUS_CANT_CONNECT or SOUP_STATUS_CANT_CONNECT_PROXY. Now that we
no longer use the message status for transport errors, we should make
soup_session_websocket_connect_async() fail with the connection error.