meson: Fix generation of gtk-doc documents and enable it in CI
It seem Gtk-doc doesn't interpret escaped parenthesis in 'ignore-decorators' as
parenthesis characters, but as grouping operators. What is happening is
that the regular expression for matching strings like 'SOUP_DEPRECATED_IN_3_0_FOR(f)'
is doing an extra matching, causing an error in parsing.

Gtk-doc has code in place for expanding an ignore-decorator expression using
parenthesis to '\(\w*\)'.