connection: rework state changes
Make soup_connection_set_state() private and the property read only. The
changes to IN_USE and IDLE states are now done by
soup_connection_set_in_use() that is called every time the connection is
assigned or unassigned to a message. When the connection is no longer
used by any message, it switches to IDLE state if it's reusable or
DISCONNECT otherwise. This simplifies the way the state is changed,
soup_connection_set_state() is now a simpler setter, it can't be called
recursively anymore. soup_connection_get_state() has been simplified
too, it was weird that a getter could also change the state, so it's now
a simpler getter that returns the current state. Callers can check if
the idle connection is still open if they need it by using
soup_connection_is_idle_open(). This removes the
SOUP_CONNECTION_REMOTE_DISCONNECTED state that is no longer used.