io-http2: ensure we stop polling the network input stream after the decoded data stream is created
In some cases we might end up polling the network input forever, if
a message read iteration handles the end of another message and there's
nothing more to read from the network. In those cases we should make
sure we stop polling the network to poll the decoded data stream
instead.