io-http2: simplify async io handling
Use a global polling source for reading and process any pending io
operation after every successful read. Every time we submit data to the
session we try to write, using a single global polling source in case
the operation would block until everything is written. This simplifies
the io handling and avoids the creation and destruction of a lot of
polling sources.