test-utils: Save Apache server root during initialization
By the time we want to use it during teardown, it is no longer safe to
call g_test_build_filename(), because some GTest state is discarded
at the end of g_test_run() (see glib#2563). As a result, the command
that should shut down the server fails, the server continues to run, and
its port is unavailable for the next test-case.

This should mostly resolve
<https://gitlab.gnome.org/GNOME/libsoup/-/issues/175>.

Signed-off-by: Simon McVittie <smcv@debian.org>