message: make it possible to force http/2 too
Now that apache supports http/2 for tests, we need a way to force http/2
because http1 is still the preferred protocol. This changes the API to
force http1 into force http version. It's not exactly the same because
force http1 also included http 1.0, but forcing http 1.1 should just
work too.