negotiate: don't print warnings if response is HTTP 200
If the server sends a success status code, then whatevs, we good. If
there's no WWW-Authenticate header, nothing more is required from us.
We're done.

If we get status code 200 and a WWW-Authenticate header that contains a
value, then go ahead and continue to try to handle it. The preexisting
code will print a warning if it's invalid, but treat is as success
anyway. The preexisting comment explains why this is necessary.

Fixes #184