server: make server clients be the connections and not the messages
Client list is now a list of connections, owned by the server. The
connection creates the messages and notifies the server with
request-started signal. The first message is created before the TLS
handshake, and then request-started is emitted to allow the user to
connect to accept-certificate signal on the message. If the connection
is persistent, the next request-started signal will be emitted if
there's an actual request from the client. This is a change in behavior,
before we always emitted the request-started when waiting for a new
request even if the request never actually started.