lib: Add names to various GSources
This helps in identifying what’s attached to a `GMainContext` when
debugging runtime problems with libsoup.

The names added in `soup-misc.c` are fairly rough; if they turn out to
not be specific enough, those methods could be changed to accept a
`const gchar *name` argument so the caller can provide a more specific
name.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>