connection-manager: don't crash if connection outlives its manager
I have no clue whether SoupConnections are expected to outlive
SoupConnectionManager or not, but it's happening, and it doesn't seem too
surprising; after all, SoupConnection is a GObject, and things can keep
references to it. Guard against this by disconnecting from the signals
of each SoupConnection when destroying the SoupConnectionManager.

Probably fixes #361