test: Avoid fatal warnings when failing ssl-test
I was trying to debug a failing ssl-test due to having TLS misconfigured.
Assertions are set to be nonfatal, but the test would also unrelatedly
abort due to overwriting a nonnull GError and unreffing a null object.
These fixes ensure that the test can run to the end even if some
assertions fail.