build config: Support using Heimdal for Windows
This adds a build option to use the Heimdal implementation of GSSAPI on Windows
for now, and does a check to ensure that we do not try to use MIT Kerberos
GSSAPI headers when building against the Heimdal libraries (or vice-versa).

Sorry, I am not aware of configuration mechanisms for Heimdal for non-Visual
Studio builds.

We still use MIT Kerberos by default.