GSSAPI: Support builds using Heimdal
After finishing the check to see whether we are using Heimdal, define a new
build-time macro that will include headers, in particular for Windows, that
will define extra things that we need for building with Heimdal.

Also add in soup-auth-negotiate.c checks that we are not attempting to mix
using Heimdal libs with MIT Kerberos headers (or vice-versa).