websocket: Add support for keepalive pong timeout
There is already support for keepalive pings on a regular interval with

    soup_websocket_connection_set_keepalive_interval (...)

Add a new function

    soup_websocket_connection_set_keepalive_pong_timeout (...)

that - if used - begins to require pings to be ponged. If a
ping does not result in a pong before the specified timeout,
the connection is assumed to be lost, and we transition to a
closed state.