meson.build: Don't check for UNIX socket support on Windows
No point for that, albeit later GLib releases added support for UNIX
sockets for Windows as later Windows 10 (and thus Windows 11) added some
support for those, so we can check for that later on whether the support
is sufficient for our needs there.

This avoids an unneeded attempt to pull in GLib for building with UNIX
support (which does not work on Windows naturally).