WIP
This fixes when 64bit values get used as 32bit values.

When possible I've changed our type to the 64bit value.

In cases where our API returns the wrong type I assert if it is too large.

In the majority of cases we simply have to use a int as it is public API
of glib. In all other cases the value gets clamped to the min or max int
value. All of these cases seem extremely unlikely in the first place.