http2: Improve management of window sizes
Currently we allow nghttp2 to manage window sizes automatically,
what this means is when we read data from the network nghttp2 immediately
counts this data as read and will request that we want more.

However libsoup is an intermediate for other applications and the
application may not have read this data yet, leaving a buffer in
libsoup.

By disabling the automatic management and only considering data
consumed when the application reads from our buffer it will
maintain a more reasonable size.

Closes #406