server: fix listening on localhost when IPv6 is not supported
If the application requests that we listen on both IPv4 and IPv6
localhost, and IPv4 succeeds but IPv6 fails, that is perfectly fine. One
or the other is good enough unless the application has specified
otherwise. We already have a case to test for this, but it's not working
because it's expecting a different error code. Maybe the error code
changed at some point.

Fixes #172