build: Switch from krb5-config to pkg-config for gssapi
This is more simple and we already depend on pkg-config heavily.

This fixed linker errors on my system though the reason is unclear
as both output `-lgssapi_krb5` (krb5-config includes other libs that
are not actually needed). So this may be a meson issue in how
declare_dependency() works. No matter the cause it seems unnecessary.