headers: Ensure Request-Line comprises entire first line
This commit reuses one pointer 'p' to check whether the first full line
of the Request-Line equals the point at which we determined the HTTP
Version to have concluded. Therefore the Method, Path and Version are good.

Newly added Unit Tests for header parsing include both issues because
each test depends upon this code change to respond appropriately.

The appropriate HTTP Response for an unexpected end of line within the
Method or Request Target should be 400 Bad Request.

Closes: #380
Closes: #441