docs: Resolved minor old gtk-docs links in MessageHeadersIter
A few links were detected using %ObjectIdentifier instead of gi-docgen
preferred link or item identification methods. Replaced with the recognized
gtk-doc code fragment #TypeName to be consistent with other areas of this
project.

Also mentioned on the ticket were SoupPathMap, but this is not subject to
GObject Introspection at this time so those pages are not generated. Old
links were upgraded in that area as well.

Closes #455