soup-session: Fix a possible memory leak of a feature object
This had been discovered by the cookies-test.c:do_remove_feature_test(),
where the feature is removed before the SoupMessage is processed, but
still with the feature left reffed, but not referenced anywhere.

Making a list of the queued features on start and unqueue them regardless
which current features are set on the session fixes the memory leak and
makes sure only queued features are unqueued (a caller could also add new
features before the message is processed).