websockets: Sec-WebSocket-Protocol should be case sensitive
According to RFC6455 (Section 1.3), while opening a handshake:

"The server selects one or none of the acceptable protocols and echoes
that value in its handshake to indicate that it has selected that
protocol."

The client must verify the negotiated protocol matches exactly one of
the requested protocols. Until now, this comparison was done non-case
sensitive, which doesn't comply with the WebSocket spec.

Closes #466