tests: Allow range tests to check more response statuses
Currently the tests can either assert that a response to a Range request
is HTTP 206 (Partial Content) or HTTP 416 (Range Not Satisfiable).
However, many of the states resulting from parsing a Range header
actually should have a HTTP 200 (OK) response, where the Range header is
ignored due to being completely invalid. This is what Apache does when,
for example, given a Range header where the end of the range is before
the start.

Add some extra arguments to the helper functions in `range-test.c` to
allow the expected response status and expected response start and end
(of the full response) to be given by the test, and change the existing
tests to use these extra arguments.

This introduces no behavioural changes for now, but will allow following
commits to test their status 200 responses.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>